optimize parse_code_hash function use
