+++ crates/cargo-contract/src/cmd/extrinsics/instantiate.rs
@@ -83,7 +83,7 @@ pub struct InstantiateCommand {
     /// The hash of the smart contract code already uploaded to the chain.
     /// If the contract has not already been uploaded use `--wasm-path` or run the `upload` command
     /// first.
-    #[clap(long, value_parser = parse_code_hash)]
+    #[clap(long, value_parser = super::parse_code_hash)]
     code_hash: Option<<DefaultConfig as Config>::Hash>,
     /// The name of the contract constructor to call
     #[clap(name = "constructor", long, default_value = "new")]
@@ -111,17 +111,6 @@ pub struct InstantiateCommand {
     /// Export the instantiate output in JSON format.
     #[clap(long, conflicts_with = "verbose")]
     output_json: bool,
-}
-
-/// Parse a hex encoded 32 byte hash. Returns error if not exactly 32 bytes.
-fn parse_code_hash(input: &str) -> Result<<DefaultConfig as Config>::Hash> {
-    let bytes = decode_hex(input)?;
-    if bytes.len() != 32 {
-        anyhow::bail!("Code hash should be 32 bytes in length")
-    }
-    let mut arr = [0u8; 32];
-    arr.copy_from_slice(&bytes);
-    Ok(arr.into())
 }
 
 /// Parse hex encoded bytes.
@@ -535,23 +524,4 @@ enum Code {
     Upload(Vec<u8>),
     /// The code hash of an on-chain Wasm blob.
     Existing(<DefaultConfig as Config>::Hash),
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn parse_code_hash_works() {
-        // with 0x prefix
-        assert!(parse_code_hash(
-            "0xd43593c715fdd31c61141abd04a99fd6822c8558854ccde39a5684e7a56da27d"
-        )
-        .is_ok());
-        // without 0x prefix
-        assert!(parse_code_hash(
-            "d43593c715fdd31c61141abd04a99fd6822c8558854ccde39a5684e7a56da27d"
-        )
-        .is_ok())
-    }
 }
